chore: remove all references to node
