+++ dashboard/package.json
@@ -58,8 +58,5 @@
 		"typescript": "^5.5.4",
 		"vite": "^5.3.5",
 		"vite-tsconfig-paths": "^4.3.2"
-	},
-	"engines": {
-		"node": ">=18.0.0"
 	}
 }
