--- tracker/src/w3c_event.js
@@ -1,12 +1,0 @@
-// @ts-nocheck
-/**
- * @fileoverview Custom externs to suppress compiler warnings.
- * @externs
- */
-/**
- * @param {string} type
- * @param {EventListener|function(!Event):(boolean|undefined)} listener
- * @param {(boolean|!AddEventListenerOptions)=} opt_options
- * @return {undefined}
- */
-var addEventListener = function (type, listener, opt_options) {};
